Refactor getManuscriptXml to AbstractArticleCrudService
